requirements.txt
algoliasearch==2.6.3
algoliasearch-django==2.0.0
amqp==5.2.0
asgiref==3.7.2
attrs==23.2.0
billiard==4.2.0
cachetools==5.3.2
celery==5.3.6
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cron_descriptor==1.4.0
dj-database-url==2.1.0
Django==4.2.7
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==4.3.1
django-filter==23.5
django-timezone-field==6.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.27.1
google-api-core==2.14.0
google-api-python-client==2.108.0
google-auth==2.24.0
google-auth-httplib2==0.1.1
googleapis-common-protos==1.61.0
httplib2==0.22.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kombu==5.3.4
Markdown==3.5.1
MarkupSafe==2.1.3
openapi-codec==1.3.2
Pillow==10.1.0
prompt-toolkit==3.0.43
protobuf==4.25.1
psycopg2==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyJWT==2.8.0
pyparsing==3.1.1
python-crontab==3.0.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
razorpay==1.4.1
redis==5.0.1
referencing==0.32.1
requests==2.31.0
rpds-py==0.17.1
rsa==4.9
setuptools==69.0.2
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
stripe==7.10.0
sweetify==2.3.1
typing_extensions==4.9.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
uuid==1.30
vine==5.1.0
wcwidth==0.2.12
whitenoise==6.6.0
